+++ stories/Controlled/index.js
@@ -1,1 +1,2 @@
 export Basic from './Basic';
+export Informed from './Informed';